backend/backend_A/requirements.txt
# Web 框架
Flask

# MongoDB 对象文档映射器 (ODM)
mongoengine

# 处理跨域资源共享 (CORS)
Flask-Cors

# 用于密码哈希
bcrypt

# 用于生成和验证 JSON Web Tokens (JWT)
PyJWT

# 用于从 .env 文件加载环境变量，保证配置安全
python-dotenv 